Port telemetry layer
