+++ common/src/main/java/com/microsoft/identity/common/adal/internal/net/DefaultConnectionService.java
@@ -36,8 +36,8 @@ import androidx.annotation.RequiresApi;
 
 import com.microsoft.identity.common.adal.internal.PowerManagerWrapper;
 import com.microsoft.identity.common.internal.telemetry.Telemetry;
-import com.microsoft.identity.common.internal.telemetry.TelemetryEventStrings;
-import com.microsoft.identity.common.internal.telemetry.events.BaseEvent;
+import com.microsoft.identity.common.java.telemetry.TelemetryEventStrings;
+import com.microsoft.identity.common.java.telemetry.events.BaseEvent;
 
 /**
  * Default connection service check network connectivity.
