+++ common/src/main/java/com/microsoft/identity/common/internal/eststelemetry/SchemaConstants.java
@@ -22,7 +22,7 @@
 // THE SOFTWARE.
 package com.microsoft.identity.common.internal.eststelemetry;
 
-import com.microsoft.identity.common.internal.telemetry.TelemetryEventStrings;
+import com.microsoft.identity.common.java.telemetry.TelemetryEventStrings;
 
 import java.util.Arrays;
