+++ common/src/main/java/com/microsoft/identity/common/internal/telemetry/events/UiEndEvent.java
@@ -22,11 +22,11 @@
 // THE SOFTWARE.
 package com.microsoft.identity.common.internal.telemetry.events;
 
-import static com.microsoft.identity.common.internal.telemetry.TelemetryEventStrings.Event;
-import static com.microsoft.identity.common.internal.telemetry.TelemetryEventStrings.EventType;
-import static com.microsoft.identity.common.internal.telemetry.TelemetryEventStrings.Key;
+import static com.microsoft.identity.common.java.telemetry.TelemetryEventStrings.Event;
+import static com.microsoft.identity.common.java.telemetry.TelemetryEventStrings.EventType;
+import static com.microsoft.identity.common.java.telemetry.TelemetryEventStrings.Key;
 
-public class UiEndEvent extends BaseEvent {
+public class UiEndEvent extends com.microsoft.identity.common.java.telemetry.events.BaseEvent {
     public UiEndEvent() {
         super();
         names(Event.UI_END_EVENT);
