+++ common/src/main/java/com/microsoft/identity/common/internal/telemetry/observers/ITelemetryAggregatedObserver.java
@@ -22,13 +22,9 @@
 // THE SOFTWARE.
 package com.microsoft.identity.common.internal.telemetry.observers;
 
-import java.util.Map;
-
 /**
- * Telemetry observer interface for MATS (Microsoft Authentication Telemetry Service) format.
- * The calling application need to implement the interface for further data processing.
- */
-public interface ITelemetryAggregatedObserver extends ITelemetryObserver<Map<String, String>> {
-    @Override
-    void onReceived(Map<String, String> telemetryData);
+ * Deprecated. Use {@link com.microsoft.identity.common.java.telemetry.observers.ITelemetryAggregatedObserver} instead.
+ **/
+@Deprecated
+public interface ITelemetryAggregatedObserver extends com.microsoft.identity.common.java.telemetry.observers.ITelemetryAggregatedObserver {
 }
