+++ common/src/main/java/com/microsoft/identity/common/internal/telemetry/observers/ITelemetryObserver.java
@@ -23,14 +23,8 @@
 package com.microsoft.identity.common.internal.telemetry.observers;
 
 /**
- * The interface function for apps to override if they want to get the Telemetry.
- */
-public interface ITelemetryObserver<T> {
-
-    /**
-     * Invoked when telemetry data is received.
-     *
-     * @param telemetryData telemetry data
-     */
-    void onReceived(T telemetryData);
+ * Deprecated. Use {@link com.microsoft.identity.common.java.telemetry.observers.ITelemetryObserver} instead.
+ **/
+@Deprecated
+public interface ITelemetryObserver<T> extends com.microsoft.identity.common.java.telemetry.observers.ITelemetryObserver<T> {
 }
