+++ common4j/src/main/com/microsoft/identity/common/java/telemetry/TelemetryPropertiesCache.java
@@ -20,14 +20,11 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.common.internal.telemetry;
+package com.microsoft.identity.common.java.telemetry;
 
-import android.content.Context;
-import android.text.TextUtils;
-
-import androidx.annotation.NonNull;
-
-import com.microsoft.identity.common.internal.cache.SharedPreferencesFileManager;
+import com.microsoft.identity.common.java.interfaces.IKeyPairStorage;
+import com.microsoft.identity.common.java.util.StringUtil;
+import lombok.NonNull;
 
 import java.util.UUID;
 
@@ -36,8 +33,6 @@ import java.util.UUID;
  */
 class TelemetryPropertiesCache {
 
-    private static final String SHARED_PREFS_NAME = "com.microsoft.common.telemetry-properties";
-
     // region Cached Properties
     /**
      * The randomly generated identifier for this device.
@@ -45,10 +40,10 @@ class TelemetryPropertiesCache {
     private static final String DEVICE_ID_GUID = "device_id_guid";
     // endregion
 
-    private final SharedPreferencesFileManager mSharedPrefs;
+    private final IKeyPairStorage mStorage;
 
-    TelemetryPropertiesCache(@NonNull final Context context) {
-        mSharedPrefs = SharedPreferencesFileManager.getSharedPreferences(context, SHARED_PREFS_NAME, -1, null);
+    TelemetryPropertiesCache(@NonNull final IKeyPairStorage storage) {
+        mStorage = storage;
     }
 
     /**
@@ -57,11 +52,11 @@ class TelemetryPropertiesCache {
      * @return The String ID used to refer to this device.
      */
     synchronized String getOrCreateRandomStableDeviceId() {
-        String deviceId = mSharedPrefs.getString(DEVICE_ID_GUID);
+        String deviceId = mStorage.get(DEVICE_ID_GUID);
 
-        if (TextUtils.isEmpty(deviceId)) {
+        if (StringUtil.isNullOrEmpty(deviceId)) {
             deviceId = UUID.randomUUID().toString();
-            mSharedPrefs.putString(DEVICE_ID_GUID, deviceId);
+            mStorage.put(DEVICE_ID_GUID, deviceId);
         }
 
         return deviceId;
